% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationFromPeakTable.parallel.R
\name{annotationFromPeakTable.parallel}
\alias{annotationFromPeakTable.parallel}
\title{Title}
\usage{
annotationFromPeakTable.parallel(
  peakTable,
  mgfFile,
  database,
  ionMode,
  CE = "all",
  tRCalibration = F,
  is.tR.file = NA,
  MS1DeltaMZ,
  MS1DeltaTR,
  MS2.sn.threshold = 3,
  MS2.noise.intensity = "minimum",
  MS2.missing.value.padding = "minimal.value",
  ms2Mode = "ida",
  diaMethod = "NA",
  MS1MS2DeltaTR,
  MS1MS2DeltaMZ,
  MS2DeltaMZ,
  scoreMode = "average",
  cores
)
}
\arguments{
\item{peakTable}{Peak table containing m/z and tr.}

\item{mgfFile}{The MGF file containing MS/MS information.}

\item{database}{The database used for identification.}

\item{ionMode}{Positive ion mode is 'P', negative ion mode is 'N'.}

\item{CE}{Collision energy，default value is 'all'.}

\item{tRCalibration}{T or F, T will do retention time calibration and F will not.}

\item{is.tR.file}{The retention time of internal standards which saved in a xlsx file.}

\item{MS1DeltaMZ}{Delta m/z of MS1.}

\item{MS1DeltaTR}{Delta retention time.}

\item{MS2.sn.threshold}{The S/N threshold of MS2.}

\item{MS2.noise.intensity}{The intensity of noise of MS2 spectrum.}

\item{MS2.missing.value.padding}{The MS2 missing value padding method, "half" or "minimal.value"}

\item{ms2Mode}{MS2 acquisition mode, 'ida' or 'dia', the default value is 'ida'}

\item{diaMethod}{If MS2 acquisition mode is "dia", a file save dia method should be provided.}

\item{MS1MS2DeltaTR}{Delta retention time between MS1 and MS2.}

\item{MS1MS2DeltaMZ}{Delta m/z between MS1 and MS2.}

\item{MS2DeltaMZ}{Delta m/z between experimental MS2 and reference MS2.}

\item{scoreMode}{The MS2 score mode, default is "average".}

\item{cores}{The number of CPU cores when computing.}
}
\value{
annotationFromPeakTableRes
}
\description{
Title
}
\examples{
annotationFromPeakTableRes <- annotationFromPeakTable.parallel(
     peakTable = system.file("extdata/peakTable","example.csv", package = "MetEx"),
     mgfFile = system.file("extdata/mgf","example.mgf", package = "MetEx"),
     database = system.file("extdata/database","MetExDB_KEGG.xlsx", package = "MetEx"),
     ionMode = "P",
     MS1DeltaMZ = 0.01,
     MS1DeltaTR = 120,
     MS1MS2DeltaTR = 5,
     MS1MS2DeltaMZ = 0.01,
     MS2DeltaMZ = 0.02,
     cores = 4)
}
