% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMS2Score.optimized.R
\name{batchMS2Score.optimized}
\alias{batchMS2Score.optimized}
\title{Title batchMS2Score.optimized}
\usage{
batchMS2Score.optimized(
  ms1Info,
  MS1MS2DeltaMZ,
  MS1MS2DeltaTR,
  mgfFilePath,
  MS2DeltaMZ,
  NeedCleanSpectra = T,
  MS2NoiseRemoval = 0.01,
  onlyKeepMax = T,
  minScore = 0.5,
  KeepNotMatched = T,
  cores = 4
)
}
\arguments{
\item{ms1Info}{the result of extracResFliter}

\item{MS1MS2DeltaMZ}{the difference of m/z between ms1Info and mgf files}

\item{MS1MS2DeltaTR}{the difference of m/z between mgf and database of MS2}

\item{cores}{The number of CPU cores when computing.}

\item{mgfFile}{mgf file path}
}
\value{
resultList
}
\description{
Title batchMS2Score.optimized
}
