% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMS2Score.parallel.R
\name{batchMS2Score.parallel}
\alias{batchMS2Score.parallel}
\title{Title batchMS2Score.parallel}
\usage{
batchMS2Score.parallel(
  ms1Info,
  ms1DeltaMZ,
  ms2DeltaMZ,
  deltaTR,
  mgfMatrix,
  mgfData,
  MS2.sn.threshold = 3,
  MS2.noise.intensity = "minimum",
  MS2.missing.value.padding = "minimal.value",
  ms2Mode = "ida",
  scoreMode = "average",
  diaMethod = "NA",
  cores = 4
)
}
\arguments{
\item{ms1Info}{the result of extracResFliter}

\item{ms1DeltaMZ}{the difference of m/z between ms1Info and mgf files}

\item{ms2DeltaMZ}{the difference of m/z between mgf and database of MS2}

\item{deltaTR}{the difference of retention time between ms1Info and mgf}

\item{mgfMatrix}{the mgf mateix that created by importMgf}

\item{mgfData}{the mgf mateix that created by importMgf}

\item{MS2.sn.threshold}{The S/N threshold of MS2.}

\item{MS2.noise.intensity}{The intensity of noise of MS2 spectrum.}

\item{MS2.missing.value.padding}{The MS2 missing value padding method, "half" or "minimal.value"}

\item{ms2Mode}{MS2 acquisition mode, 'ida' or 'dia', the default value is 'ida'}

\item{scoreMode}{The MS2 score mode, default is "average".}

\item{diaMethod}{If MS2 acquisition mode is "dia", a file save dia method should be provided.}

\item{cores}{The number of CPU cores when computing.}
}
\value{
ms2ScoreResult
}
\description{
Title batchMS2Score.parallel
}
\examples{
load(system.file("extdata/testData", "ms1InfoTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "mgfMatrixTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "mgfDataTest.rda", package = "MetEx"))
ms2ScoreResult <- batchMS2Score.parallel(ms1Info = ms1InfoTest,
                                         ms1DeltaMZ = 0.01,
                                         ms2DeltaMZ = 0.02,
                                         deltaTR = 15,
                                         mgfMatrix = mgfMatrixTest,
                                         mgfData = mgfDataTest,
                                         cores = 1)
}
\references{
Tsugawa, Hiroshi , et al. "MS-DIAL: data-independent MS/MS deconvolution for comprehensive metabolome analysis." Nature Methods 12.6(2015):523-526.
}
