% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMS2Score.QQQ.R
\name{batchMS2Score.QQQ}
\alias{batchMS2Score.QQQ}
\title{Title batchMS2Score.QQQ}
\usage{
batchMS2Score.QQQ(
  ms1Info,
  ms1DeltaMZ,
  ms2DeltaMZ,
  deltaTR,
  intThre,
  mgfMatrix,
  mgfData
)
}
\arguments{
\item{ms1Info}{the result of extracResFliter}

\item{ms1DeltaMZ}{the difference of m/z between ms1Info and mgf files}

\item{ms2DeltaMZ}{the difference of m/z between mgf and database of MS2}

\item{deltaTR}{the difference of retention time between ms1Info and mgf}

\item{intThre}{Intensity threshold}

\item{mgfMatrix}{the mgf mateix that created by importMgf}

\item{mgfData}{the mgf data that created by inportMgf}
}
\value{
ms2ScoreResult
}
\description{
Title batchMS2Score.QQQ
}
\examples{
load(system.file("extdata/testData", "ms1InfoQQQTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "mgfMatrixTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "mgfDataTest.rda", package = "MetEx"))
ms2ScoreResult <- batchMS2Score.QQQ(ms1Info = ms1InfoQQQTest,
                                    ms1DeltaMZ = 0.01,
                                    ms2DeltaMZ = 0.02,
                                    deltaTR = 15,
                                    intThre = 270,
                                    mgfMatrix = mgfMatrixTest,
                                    mgfData = mgfDataTest)
}
