% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMS2Score.R
\name{batchMS2Score}
\alias{batchMS2Score}
\title{Title batchMS2Score}
\usage{
batchMS2Score(
  ms1Info,
  ms1DeltaMZ,
  ms2DeltaMZ,
  deltaTR,
  mgfMatrix,
  mgfData,
  MS2.sn.threshold = 3,
  MS2.noise.intensity = "minimum",
  MS2.missing.value.padding = "half",
  ms2Mode = "ida",
  scoreMode = "average",
  diaMethod = "NA"
)
}
\arguments{
\item{ms1Info}{the result of extracResFliter}

\item{ms1DeltaMZ}{the difference of m/z between ms1Info and mgf files}

\item{ms2DeltaMZ}{the difference of m/z between mgf and database of MS2}

\item{deltaTR}{the difference of retention time between ms1Info and mgf}

\item{mgfMatrix}{the mgf mateix that created by importMgf}

\item{mgfData}{the mgf data that created by inportMgf}

\item{MS2.sn.threshold}{The S/N threshold of MS2.}

\item{MS2.noise.intensity}{The intensity of noise of MS2 spectrum.}

\item{MS2.missing.value.padding}{The MS2 missing value padding method, "half" or "minimal.value"}

\item{ms2Mode}{MS2 acquisition mode, there are two selection for this parameter, 'ida' and 'dia'}

\item{scoreMode}{MS2 score mode, 'obverse','reverse' and 'average'}

\item{diaMethod}{if the MS2 acquisition mode is dia, a diaMethod is necessary. If is ida, diaMethod can be set as "NA"}
}
\value{
ms2ScoreResult
}
\description{
Title batchMS2Score
}
\examples{
load(system.file("extdata/testData", "ms1InfoTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "mgfMatrixTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "mgfDataTest.rda", package = "MetEx"))
ms2ScoreResult <- batchMS2Score(ms1Info = ms1InfoTest,
                                ms1DeltaMZ = 0.01,
                                ms2DeltaMZ = 0.02,
                                deltaTR = 15,
                                mgfMatrix = mgfMatrixTest,
                                mgfData = mgfDataTest)
}
\references{
Tsugawa, Hiroshi , et al. "MS-DIAL: data-independent MS/MS deconvolution for comprehensive metabolome analysis." Nature Methods 12.6(2015):523-526.
}
