% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassicalAnnotation.R
\name{ClassicalAnnotation}
\alias{ClassicalAnnotation}
\title{Title ClassicalAnnotation}
\usage{
ClassicalAnnotation(
  peakTable,
  mgfFilePath,
  database,
  ionMode,
  CE = "all",
  tRCalibration = F,
  is.tR.file = NA,
  MS1DeltaMZ,
  MS1DeltaTR,
  MS1MS2DeltaTR,
  MS1MS2DeltaMZ,
  MS2DeltaMZ,
  NeedCleanSpectra = T,
  MS2NoiseRemoval = 0.01,
  onlyKeepMax = T,
  minScore = 0.5,
  KeepNotMatched = T,
  cores = 1
)
}
\arguments{
\item{peakTable}{Peak table containing m/z and tr.}

\item{database}{The database used for identification.}

\item{ionMode}{Positive ion mode is 'P', negative ion mode is 'N'.}

\item{CE}{Collision energy，default value is 'all'.}

\item{tRCalibration}{T or F, T will do retention time calibration and F will not.}

\item{is.tR.file}{The retention time of internal standards which saved in a xlsx file.}

\item{MS1DeltaMZ}{Delta m/z of MS1.}

\item{MS1DeltaTR}{Delta retention time.}

\item{MS1MS2DeltaTR}{Delta retention time between MS1 and MS2.}

\item{MS1MS2DeltaMZ}{Delta m/z between MS1 and MS2.}

\item{MS2DeltaMZ}{Delta m/z between experimental MS2 and reference MS2.}

\item{NeedCleanSpectra}{T}

\item{MS2NoiseRemoval}{0.01}

\item{onlyKeepMax}{T}

\item{minScore}{0.5}

\item{KeepNotMatched}{T}

\item{cores}{The number of CPU cores when computing.}

\item{mgfFile}{The MGF file containing MS/MS information.}
}
\value{
ClassicalAnnotationRes
}
\description{
Title ClassicalAnnotation
}
