% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg.R
\name{emgGaussian_guess}
\alias{emgGaussian_guess}
\title{Guess function for initial exponentially modified gaussian
parameters and bounds}
\usage{
emgGaussian_guess(x, y)
}
\arguments{
\item{x}{(numeric) x values (e.g. retention time)}

\item{y}{(numeric) y observed values (e.g. spectra intensity)}
}
\value{
A list of guessed starting parameters \code{list()$init_params},
lower \code{list()$lower_bounds} and upper bounds \code{list()$upper_bounds}
(\code{$gamma}, \code{$center}, \code{$sigma}, \code{$amplitude})
}
\description{
Guess function for initial exponentially modified gaussian parameters and bounds, at the moment only checks the x position
The functions come from peakPantheR, because they cannot be used alone for the time being, we temporarily put them in MetEx and marked the references.
}
