% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg.R
\name{emgGaussian_minpack.lm}
\alias{emgGaussian_minpack.lm}
\title{Implementation of the Exponentially Modified Gaussian (EMG) peak
shape for use with minpack.lm}
\usage{
emgGaussian_minpack.lm(params, xx)
}
\arguments{
\item{params}{(list) exponentiall modified gaussian parameters
(\code{params$gamma}, \code{params$center}, \code{params$sigma},
\code{params$amplitude})}

\item{xx}{(numeric) values at which to evalute the exponentially
modified gaussian}
}
\value{
value of the exponentially modified gaussian evaluated at xx
}
\description{
Implementation of the  Exponentially Modified Gaussian (EMG) peak shape for use with minpack.lm
The functions come from peakPantheR, because they cannot be used alone for the time being, we temporarily put them in MetEx and marked the references.
}
