% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg.R
\name{emgGaussian_minpack.lm_objectiveFun}
\alias{emgGaussian_minpack.lm_objectiveFun}
\title{Exponentially Modified Gaussian minpack.lm objective function}
\usage{
emgGaussian_minpack.lm_objectiveFun(params, observed, xx)
}
\arguments{
\item{params}{(list) exponentially modified gaussian parameters
(\code{params$gamma}, \code{params$center}, \code{params$sigma},
\code{params$amplitude})}

\item{observed}{(numeric) observed y value at xx}

\item{xx}{(numeric) value at which to evalute the exponentially
modified gaussian}
}
\value{
difference between observed and expected exponentially modified
gaussian value evaluated at xx
}
\description{
Exponentially Modified Gaussian (EMG) minpack.lm objective function, calculates residuals using the EMG Peak Shape
The functions come from peakPantheR, because they cannot be used alone for the time being, we temporarily put them in MetEx and marked the references.
}
