% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactMsp.R
\name{exactMsp}
\alias{exactMsp}
\title{Title
exactMsp}
\usage{
exactMsp(mspData)
}
\arguments{
\item{mspData}{the msp data.}
}
\value{
mspDataframe
}
\description{
Title
exactMsp
}
\examples{
load(system.file("extdata/testData", "mspDataTest.rda", package = "MetEx"))
mspDataframe <- exactMsp(mspDataTest)
}
