% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracResFilter.R
\name{extracResFilter}
\alias{extracResFilter}
\title{Title
Fliter the result of targetExtraction by entropy and intensity.}
\usage{
extracResFilter(
  targExtracRes,
  entroThre,
  intThre,
  classficationMethod = "NoSVM"
)
}
\arguments{
\item{targExtracRes, }{the result of targetExtraction}

\item{entroThre, }{the threshold of entropy}

\item{intThre, }{the threshold of intensity}

\item{classficationMethod}{Classfication Method, "NoSVM" or "SVM".}
}
\value{
ms1Info
}
\description{
Title
Fliter the result of targetExtraction by entropy and intensity.
}
\examples{
load(system.file("extdata/testData", "targExtracResTest.rda", package = "MetEx"))
ms1Info <- extracResFilter(targExtracRes = targExtracResTest,
                           entroThre=2,
                           intThre=250,
                           classficationMethod="NoSVM")
}
