% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg.R
\name{gaussian_cerf}
\alias{gaussian_cerf}
\title{Gaussian Error function}
\usage{
gaussian_cerf(x)
}
\arguments{
\item{x}{(numeric) value at which to evaluate the gaussian error function}
}
\value{
Value of the gaussian error function evaluated at x
}
\description{
Implementation of the gaussian error function
The functions come from peakPantheR, because they cannot be used alone for the time being, we temporarily put them in MetEx and marked the references.
}
