% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMgf.R
\name{importMgf}
\alias{importMgf}
\title{Title
Import mgf files and generate mgfData (raw file) and mgfMatrix (a matrix of MS2 data).}
\usage{
importMgf(mgfFile)
}
\arguments{
\item{mgfFile}{the mgf file.}
}
\value{
mgfList
}
\description{
Title
Import mgf files and generate mgfData (raw file) and mgfMatrix (a matrix of MS2 data).
}
\examples{
mgfList <- importMgf(mgfFile = system.file("extdata/mgf", "example.mgf", package = "MetEx"))
}
