% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intEntropyCalculator.R
\name{intEntropyCalculator}
\alias{intEntropyCalculator}
\title{Title
Calculate entropy}
\usage{
intEntropyCalculator(dataPeakSheak)
}
\arguments{
\item{dataPeakSheak, }{a matrix which has two column of MSMS}
}
\value{
entropy
}
\description{
Title
Calculate entropy
}
\examples{
load(system.file("extdata/testData", "dataPeakSheakTest.rda", package = "MetEx"))
entropy <- intEntropyCalculator(dataPeakSheak = dataPeakSheakTest)
}
\references{
Ju R, Liu X, Zheng F, et al. Removal of False Positive Features to Generate Authentic Peak Table for High-resolution Mass Spectrometry-based Metabolomics Study[J]. Analytica Chimica Acta, 2019.
}
