% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonReader.R
\name{jsonReader}
\alias{jsonReader}
\title{Title
Convert json file to csv file.}
\usage{
jsonReader(jsonFile, csvFile, pagesize = 100)
}
\arguments{
\item{jsonFile}{json file}

\item{csvFile}{csv file}

\item{pagesize}{number of lines to read/write from/to the connection per iteration.}
}
\description{
Title
Convert json file to csv file.
}
\examples{
json.df <- jsonReader(jsonFile =
                      system.file("extdata/json", "example.json", package = "MetEx"),
                      csvFile = "E:/test.csv")
}
