% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGdownloader.R
\name{KEGGdownloader}
\alias{KEGGdownloader}
\title{Title
Download compounds in KEGG.}
\usage{
KEGGdownloader(items = "all")
}
\arguments{
\item{items}{only for test.}
}
\value{
kegg.compound.database.df
}
\description{
Title
Download compounds in KEGG.
}
\examples{
kegg.compound.database.df <- KEGGdownloader(items = 10)
}
