% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2dataframe.R
\name{list2dataframe}
\alias{list2dataframe}
\title{Title
Tranfer list to dataframe.}
\usage{
list2dataframe(ms2List)
}
\arguments{
\item{ms2List}{a list containing the information of MSMS}
}
\value{
ms2Dataframe
}
\description{
Title
Tranfer list to dataframe.
}
\examples{
load(system.file("extdata/testData", "ms2ListTest.rda", package = "MetEx"))
ms2Dataframe <- list2dataframe(ms2ListTest)
}
