% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMetExResult.R
\name{mergeMetExResult}
\alias{mergeMetExResult}
\title{Title mergeMetExResult}
\usage{
mergeMetExResult(resultPath)
}
\arguments{
\item{resultPath}{MetEx results path}
}
\value{
resultList
}
\description{
Title mergeMetExResult
}
