% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergems2.R
\name{mergems2}
\alias{mergems2}
\title{Title
merge MSMS}
\usage{
mergems2(MS2.matrix)
}
\arguments{
\item{MS2.matrix, }{a matrix containing two column}
}
\value{
ms2
}
\description{
Title
merge MSMS
}
\examples{
load(system.file("extdata/testData", "MS2MatrixTest.rda", package = "MetEx"))
ms2 <- mergems2(MS2.matrix = MS2MatrixTest)
}
