% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms1ms2Match.R
\name{ms1ms2Match}
\alias{ms1ms2Match}
\title{Title
When mz and rt are given, find the matched MSMS.}
\usage{
ms1ms2Match(
  mz,
  tr,
  ms1DeltaMZ,
  deltaTR,
  mgfMatrix,
  mgfData,
  ms2Mode,
  diaMethod = "NA"
)
}
\arguments{
\item{mz, }{m/z}

\item{tr, }{retention time}

\item{ms1DeltaMZ, }{the tolerence of m/z}

\item{deltaTR, }{the tolerence of retention time}

\item{mgfMatrix, }{a matrix containing the information of mgf}

\item{mgfData, }{a data containing the information of mgf}

\item{ms2Mode, }{the mode of MS2 acquisition, ida or dia}

\item{diaMethod, }{if the ms2Mode is dia, a txt file containing dia method.}
}
\value{
matchedMS2
}
\description{
Title
When mz and rt are given, find the matched MSMS.
}
\examples{
load(system.file("extdata/testData", "mgfMatrixTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "mgfDataTest.rda", package = "MetEx"))
matchedMS2 <- ms1ms2Match(mz = 118.0106,
                          tr = 120,
                          ms1DeltaMZ = 0.01,
                          deltaTR = 12,
                          mgfMatrix = mgfMatrixTest,
                          mgfData = mgfDataTest,
                          ms2Mode = "ida",
                          diaMethod = "NA")
}
