% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2Score.old.R
\name{ms2Score.old}
\alias{ms2Score.old}
\title{Title
Give MSMS Score.}
\usage{
ms2Score.old(ms2Act, ms2DB, ms2DeltaMZ, sn.threshold = 3, scoreMode)
}
\arguments{
\item{ms2Act, }{measured ms2}

\item{ms2DB, }{ms2 in database}

\item{ms2DeltaMZ, }{the tolerence between ms2Act and ms2DB}

\item{sn.threshold}{The S/N threshold of MS2 spectrum.}

\item{scoreMode, }{the mode of score, obverse, reverse or average}
}
\value{
scoreOfMS2
}
\description{
Title
Give MSMS Score.
}
\examples{
load(system.file("extdata/testData", "ms2ActTest.rda", package = "MetEx"))
load(system.file("extdata/testData", "ms2DBTest.rda", package = "MetEx"))
scoreOfMS2 <- ms2Score.old(ms2Act = ms2ActTest,
                           ms2DB = ms2DBTest,
                           ms2DeltaMZ=0.02,
                           sn.threshold = 3,
                           scoreMode='average')
}
