% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2ScoreSpectralEntropy.R
\name{ms2ScoreSpectralEntropy}
\alias{ms2ScoreSpectralEntropy}
\title{Title
Give MSMS Score using spectral entropy.}
\usage{
ms2ScoreSpectralEntropy(
  spec_query,
  spec_reference,
  ms2_da,
  need_clean_spectra = TRUE,
  noise_removal = 0.01
)
}
\arguments{
\item{spec_query, }{measured ms2}

\item{spec_reference, }{ms2 in database}

\item{ms2_da, }{the tolerence between ms2Act and ms2DB}
}
\value{
scoreOfMS2
}
\description{
Title
Give MSMS Score using spectral entropy.
}
\references{
Li, Y., Kind, T., Folz, J. et al. Spectral entropy outperforms MS/MS dot product similarity for small-molecule compound identification. Nat Methods 18, 1524–1531 (2021). https://doi.org/10.1038/s41592-021-01331-z
}
