% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDectAndEntroCal.R
\name{peakDectAndEntroCal}
\alias{peakDectAndEntroCal}
\title{Title
Use EIC data to do peak detection and calculate entropy of each peaks within a retention time range.}
\usage{
peakDectAndEntroCal(eicData, trRange, m = 200)
}
\arguments{
\item{eicData, }{a matrix containing EIC data, the first column is rt (retention time) and the second column is intensity.}

\item{trRange}{retention time range for extraction.}

\item{m}{parameter of peak detection.}
}
\value{
extractedPeaks
}
\description{
Title
Use EIC data to do peak detection and calculate entropy of each peaks within a retention time range.
}
\examples{
load(system.file("extdata/testData", "eicDataTest.rda", package = "MetEx"))
extractedPeaks <- peakDectAndEntroCal(eicData = eicDataTest, trRange=20)
}
\references{
Ju R, Liu X, Zheng F, et al. Removal of False Positive Features to Generate Authentic Peak Table for High-resolution Mass Spectrometry-based Metabolomics Study[J]. Analytica Chimica Acta, 2019.
}
