% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDetection.R
\name{peakDetection}
\alias{peakDetection}
\title{Title
Use function in python to do peak detection.}
\usage{
peakDetection(eicData, m = 200)
}
\arguments{
\item{eicData, }{a matrix containing EIC data, the first column is rt (retention time) and the second column is intensity.}

\item{m}{parameter of peak detection.}
}
\value{
extractedPeaks
}
\description{
Title
Use function in python to do peak detection.
}
\examples{
load(system.file("extdata/testData", "eicDataTest.rda", package = "MetEx"))
extractedPeaks <- peakDetection(eicData = eicDataTest, m=200)
}
\references{
https://github.com/stas-g/findPeaks
}
