% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakGroup.R
\name{peakGroup}
\alias{peakGroup}
\title{Title
Peak Group}
\usage{
peakGroup(MetExAnnotationResList)
}
\arguments{
\item{MetExAnnotationResList}{A list of MetEx annotation}
}
\value{
res.sheet
}
\description{
Title
Peak Group
}
\examples{
load(system.file("extdata/testData","MetExAnnotationResList.Rda", package = "MetEx"))
res.sheet <- peakGroup(MetExAnnotationResList)
}
