% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMsp.R
\name{readMsp}
\alias{readMsp}
\title{Title
This function is come form R package named "metaMS". The original name of this function is "read.msp"
Beacuse of this function have bug when reading msp files with some complicated compound name. So I changed
it to fit my objective. Function reads an msp file, and returns one list of compounds. Argument only.org allow
to exclude all molecular formulas with non-organic elements. What exactly are organic elements can be defined with
the optional argument org.set. A second optional argument indicates which fields should never be converted into numbers,
even if possible.}
\usage{
readMsp(
  file,
  only.org = FALSE,
  org.set = c("C", "H", "D", "N", "O", "P", "S"),
  noNumbers = NULL
)
}
\arguments{
\item{file, }{the msp file}

\item{only.org}{FALSE}

\item{org.set}{c("C", "H", "D", "N", "O", "P", "S")}

\item{noNumbers}{NULL}
}
\value{
mspData
}
\description{
Title
This function is come form R package named "metaMS". The original name of this function is "read.msp"
Beacuse of this function have bug when reading msp files with some complicated compound name. So I changed
it to fit my objective. Function reads an msp file, and returns one list of compounds. Argument only.org allow
to exclude all molecular formulas with non-organic elements. What exactly are organic elements can be defined with
the optional argument org.set. A second optional argument indicates which fields should never be converted into numbers,
even if possible.
}
\references{
This function is come form R package named "metaMS". The original name of this function is "read.msp"
}
