% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retentionTimeCalibration.R
\name{retentionTimeCalibration}
\alias{retentionTimeCalibration}
\title{Title
Use internal standard retention to cailbrate retention time of metabolites in database.}
\usage{
retentionTimeCalibration(is.tR.file, database.df)
}
\arguments{
\item{is.tR.file, }{the txt file of IS retention times in database and experimental.}

\item{database.df, }{the imported database dataframe.}
}
\value{
calibrated.RT.db
}
\description{
Title
Use internal standard retention to cailbrate retention time of metabolites in database.
}
\examples{
load(system.file("extdata/testData", "dbDataTest.rda", package = "MetEx"))
calibrated.RT.db <- retentionTimeCalibration(is.tR.file =
                    system.file("extdata/trCalibration",
                    "IS-for-tR-calibration.xlsx", package = "MetEx"),
                    database.df = dbDataTest)
}
\references{
Huan, T. et al. DnsID in MyCompoundID for rapid identification of dansylated amine- and phenol-containing metabolites in LC-MS-based metabolomics. Anal. Chem. 87, 9838-9845 (2015).
}
