% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg.R
\name{skewedGaussian_minpack.lm}
\alias{skewedGaussian_minpack.lm}
\title{Implementation of the Skewed Gaussian peak shape for use with
minpack.lm}
\usage{
skewedGaussian_minpack.lm(params, xx)
}
\arguments{
\item{params}{(list) skewed gaussian parameters (\code{params$gamma},
\code{params$center}, \code{params$sigma}, \code{params$amplitude})}

\item{xx}{(numeric) values at which to evalute the skewed gaussian}
}
\value{
value of the skewed gaussian evaluated at xx
}
\description{
Implementation of the Skewed Gaussian peak shape for use with minpack.lm.
The functions come from peakPantheR, because they cannot be used alone for the time being, we temporarily put them in MetEx and marked the references.
}
