% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emg.R
\name{skewedGaussian_minpack.lm_objectiveFun}
\alias{skewedGaussian_minpack.lm_objectiveFun}
\title{Skewed Gaussian minpack.lm objective function}
\usage{
skewedGaussian_minpack.lm_objectiveFun(params, observed, xx)
}
\arguments{
\item{params}{(list) skewed gaussian parameters (\code{params$gamma},
\code{params$center}, \code{params$sigma}, \code{params$amplitude})}

\item{observed}{(numeric) observed y value at xx}

\item{xx}{(numeric) value at which to evalute the skewed gaussian}
}
\value{
difference between observed and expected skewed gaussian value
evaluated at xx
}
\description{
Skewed Gaussian minpack.lm objective function, calculates residuals using the skewed gaussian Peak Shape
The functions come from peakPantheR, because they cannot be used alone for the time being, we temporarily put them in MetEx and marked the references.
}
