% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetExtraction.optimized.R
\name{targetExtraction.optimized}
\alias{targetExtraction.optimized}
\title{Title
Targeted extraction from untargeted metabolomics data by using the screen result of dbImporter}
\usage{
targetExtraction.optimized(
  msRawData,
  dbData,
  deltaMZ,
  deltaTR,
  trRange = 30,
  m = 200
)
}
\arguments{
\item{msRawData, }{the MS data which have a XCMS-support filetype (such as mzXML)}

\item{dbData, }{the metabolomics database which first processed by dbImporter}

\item{deltaMZ, }{the m/z tolerence of MS1 between database and raw data}

\item{deltaTR, }{the retention time tolerence of MS1 between database and raw data}

\item{trRange}{retention time range for extraction.}

\item{m}{parameter of peak detection.}
}
\value{
targExtracRes
}
\description{
Title
Targeted extraction from untargeted metabolomics data by using the screen result of dbImporter
}
\examples{
load(system.file("extdata/testData", "dbDataTest.rda", package = "MetEx"))
targExtracRes <- targetExtraction.optimized(
                 msRawData = system.file("extdata/mzXML", "example.mzXML", package = "MetEx"),
                 dbData = dbDataTest,
                 deltaMZ=0.01,
                 deltaTR=60)
}
