% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlReader.R
\name{xmlReader}
\alias{xmlReader}
\title{Title
Convert json file to csv file.}
\usage{
xmlReader(xmlFile, csvFile)
}
\arguments{
\item{xmlFile}{xml file}

\item{csvFile}{csv file}
}
\description{
Title
Convert json file to csv file.
}
\examples{
xml.df <- xmlReader(xmlFile =
                    system.file("extdata/xml", "example.xml", package = "MetEx"),
                    csvFile = "E:/test.csv")
}
