% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiedResFilter.R
\name{identifiedResFilter}
\alias{identifiedResFilter}
\title{Title
Filter identified Result}
\usage{
identifiedResFilter(batchMS2ScoreResult, MS2score)
}
\arguments{
\item{batchMS2ScoreResult}{batchMS2ScoreResult.}

\item{MS2score}{The threshold of MS2 score.}
}
\value{
sheets
}
\description{
Title
Filter identified Result
}
\examples{
load(system.file("extdata/testData", "batchMS2ScoreResult.Rda", package = "MetEx"))
filteredres <- identifiedResFilter(batchMS2ScoreResult = batchMS2ScoreResult, MS2score = 0.6)
}
