% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetExAnnotation.R
\name{MetExAnnotation}
\alias{MetExAnnotation}
\title{Title
Annotate metabolites by MetEx in one line.}
\usage{
MetExAnnotation(
  dbFile,
  ionMode,
  CE = "all",
  tRCalibration = F,
  is.tR.file = NA,
  msRawData,
  MS1deltaMZ,
  MS1deltaTR,
  trRange = 30,
  m = 200,
  entroThre,
  intThre,
  classficationMethod = "NoSVM",
  mgfFilePath,
  MS1MS2DeltaMZ,
  MS1MS2DeltaTR,
  MS2DeltaMZ,
  NeedCleanSpectra = T,
  MS2NoiseRemoval = 0.01,
  onlyKeepMax = T,
  minScore = 0.5,
  KeepNotMatched = T,
  MS2scoreFilter,
  cores = 1
)
}
\arguments{
\item{dbFile}{The file path of database.}

\item{ionMode}{The ion mode of experiment, "P" or "N".}

\item{CE}{Collision energy, the default value is "all".}

\item{tRCalibration}{Whether to perform retention time calibration.}

\item{is.tR.file}{If retention time calibration is performed, the retention time of IS saved in xlsx should be imported.}

\item{msRawData}{The file path of mzXML file.}

\item{MS1deltaMZ}{Tolerance of MS1.}

\item{MS1deltaTR}{Tolerance of retention time between experiment and database.}

\item{entroThre}{The information entropy threshold.}

\item{intThre}{The peak height threshold.}

\item{classficationMethod}{The classficatioin method of true signals and noises, "NoSVM" or "SVM" is avilable and "NoSVM" is the default value.}

\item{MS1MS2DeltaMZ}{The tolerance of MS between MS1 and MS2.}

\item{MS1MS2DeltaTR}{The tolerance of retention time between MS1 and MS2.}

\item{MS2DeltaMZ}{The tolerance of MS between experiment MS2 and database MS2.}

\item{NeedCleanSpectra}{T}

\item{MS2NoiseRemoval}{0.01}

\item{onlyKeepMax}{T}

\item{minScore}{0.5}

\item{KeepNotMatched}{T}

\item{MS2scoreFilter}{The threshold of MS2 score.}

\item{cores}{The CPU cores used for parallel computation.}

\item{mgfFile}{The file path of mgfFile.}
}
\value{
MetExAnnotationRes
}
\description{
Title
Annotate metabolites by MetEx in one line.
}
